Added interface IState.
Initially, it was thought that this library should be able to handle
primitive types in .NET. However, this has proven to be difficult. The
API and the internals of the library has become a huge mess because of
this design decision. Therefore, it is decided that this library should
move onto the direction of encapsulating away the hosted data types by
means of introducing this IStore interface. Hosted data types are
expected to implement the IStore interface.

One downside of this would be the difficulty of creating granular
reducers that would reduce primitive values. We expect to introduce few
helper functions for this purpose. These helpers would wrap primitives
in IStore objects, which would then allow the consumers to use
Redux.NET as if it supported premitive types.